dipHUD 1.1 - The Requests Version
In this version, the colors for the health, ammo, and medic charge have
been converted to an easy to edit version. To change the colors,
navigate to ClientScheme.res in the resource folder, and change the RGBA
values of the custom colors. Also, I added fog's crosshairs as well as
the fighter pilot crosshair with support for 4:3, 16:9 and 16:10.